Make App simple again
